+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# v0.2.0
+
+BREAKING CHANGES:
+
+* resource/time_static: The `expiration_` arguments have been moved to the new `time_rotating` resource as `rotation_` arguments.
+
+FEATURES:
+
+* **New Resource:** `time_rotating`
+
 # v0.1.0
 
 FEATURES: